Refactor checkbox, elipsis and description event handlers
