+++ src/taskStore.js
@@ -1,3 +1,5 @@
+import Helper from "./helpers.js";
+
 class TaskStore {
   constructor() {
     this.tasks = JSON.parse(localStorage.getItem('taskstore')) || [];
@@ -62,55 +64,11 @@ class TaskStore {
       </li>`;
     });
 
-    if (this.getTasks().length > 0) {
-      const checkBox = document.querySelectorAll('input[type=checkbox]');
-      checkBox.forEach((box) => {
-        box.addEventListener('change', (event) => {
-          const task = event.target.parentNode.querySelector('.desc');
-          task.classList.toggle('line-through');
-          this.setTaskStaus(parseInt(event.target.id) - 1);
-        });
-      });
-
-      const btnDelete = document.querySelectorAll('.btnDelete');
-      btnDelete.forEach((btn) => {
-        btn.addEventListener('click', (event) => {
-          this.removeTask(false, parseInt(event.target.getAttribute('data-index'), 10));
-        });
-      });
-
-      const editFields = document.querySelectorAll('.desc');
-
-      editFields.forEach((editField) => editField.addEventListener('keypress', (event) => {
-        if (event.key === 'Enter' && event.target.value !== '') {
-          event.preventDefault();
-          const id = parseInt(event.target.getAttribute('data-pos'), 10);
-          const data = event.target.value;
-          this.editTask({ id, data });
-          event.target.disabled = true;
-          event.target.parentNode.querySelector('.edit-desc').classList.toggle('d-none');
-        }
-      }));
-
-      const textFields = document.querySelectorAll('.edit-desc');
-
-      textFields.forEach((editField) => {
-        editField.addEventListener('click', (event) => {
-          editFields.forEach((editField) => {
-            editField.disabled = true;
-            editField.parentNode.classList.remove('high-light');
-            editField.parentNode.querySelector('.edit-desc').classList.remove('d-none');
-            editField.parentNode.querySelector('.fa-trash').classList.add('d-none');
-          });
-          event.target.parentNode.classList.add('high-light');
-          const task = event.target.parentNode.querySelector('.desc');
-          const deleteButton = event.target.parentNode.querySelector('.fa-trash');
-          task.disabled = false;
-          task.focus();
-          event.target.classList.toggle('d-none');
-          deleteButton.classList.toggle('d-none');
-        });
-      });
+    if (this.getTasks().length > 0) { 
+      Helper.checkBoxHandler(this);
+      Helper.deleteButtonsHandler(this);
+      Helper.editElipsisHandler(this);
+      Helper.editTextHandler(this);
     }
   }
